refs #151 Translate at preferences
