+++ src/constants/visibility.js
@@ -1,18 +1,20 @@
+import i18n from '../config/i18n'
+
 export default {
   Public: {
-    name: 'public',
+    name: i18n.t('preferences.general.visibility.public'),
     value: 0
   },
   Unlisted: {
-    name: 'unlisted',
+    name: i18n.t('preferences.general.visibility.unlisted'),
     value: 1
   },
   Private: {
-    name: 'private',
+    name: i18n.t('preferences.general.visibility.private'),
     value: 2
   },
   Direct: {
-    name: 'direct',
+    name: i18n.t('preferences.general.visibility.direct'),
     value: 3
   }
 }
