+++ src/main/index.js
@@ -9,13 +9,13 @@ import simplayer from 'simplayer'
 import path from 'path'
 import ContextMenu from 'electron-context-menu'
 import * as Splashscreen from '@trodi/electron-splashscreen'
+import openAboutWindow from 'about-window'
 
 import Authentication from './auth'
 import Account from './account'
 import Streaming from './streaming'
 import Preferences from './preferences'
 import Hashtags from './hashtags'
-import ApplicationMenu from './menu'
 import i18n from '../config/i18n'
 
 /**
@@ -113,7 +113,7 @@ async function createWindow () {
   /**
    * Set application menu
    */
-  ApplicationMenu(mainWindow, accountsChange, i18n)
+  ApplicationMenu(accountsChange, i18n)
 
   /**
    * Set dock menu for mac
@@ -646,3 +646,166 @@ app.on('ready', () => {
  */
 
 class EmptyTokenError {}
+
+/**
+ * Set application menu
+ */
+const ApplicationMenu = (accountsChange, i18n) => {
+  /**
+   * For mac menu
+   */
+  const macGeneralMenu = process.platform !== 'darwin' ? [] : [
+    {
+      type: 'separator'
+    },
+    {
+      label: i18n.t('main_menu.application.services'),
+      role: 'services',
+      submenu: []
+    },
+    {
+      type: 'separator'
+    },
+    {
+      label: i18n.t('main_menu.application.hide'),
+      role: 'hide'
+    },
+    {
+      label: i18n.t('main_menu.application.hide_others'),
+      role: 'hideothers'
+    },
+    {
+      label: i18n.t('main_menu.application.show_all'),
+      role: 'unhide'
+    }
+  ]
+
+  const template = [
+    {
+      label: i18n.t('main_menu.application.name'),
+      submenu: [
+        {
+          label: i18n.t('main_menu.application.about'),
+          role: 'about',
+          click: () => {
+            openAboutWindow({
+              icon_path: path.resolve(__dirname, '../../build/icons/256x256.png'),
+              copyright: 'Copyright (c) 2018 AkiraFukushima',
+              package_json_dir: path.resolve(__dirname, '../../'),
+              open_devtools: process.env.NODE_ENV !== 'production'
+            })
+          }
+        },
+        {
+          type: 'separator'
+        },
+        {
+          label: i18n.t('main_menu.application.preferences'),
+          accelerator: 'CmdOrCtrl+,',
+          click: () => {
+            mainWindow.webContents.send('open-preferences')
+          }
+        },
+        ...macGeneralMenu,
+        {
+          type: 'separator'
+        },
+        {
+          label: i18n.t('main_menu.application.quit'),
+          accelerator: 'CmdOrCtrl+Q',
+          role: 'quit'
+        }
+      ]
+    },
+    {
+      label: i18n.t('main_menu.toot.name'),
+      submenu: [
+        {
+          label: i18n.t('main_menu.toot.new'),
+          accelerator: 'CmdOrCtrl+N',
+          click: () => {
+            mainWindow.webContents.send('CmdOrCtrl+N')
+          }
+        }
+      ]
+    },
+    {
+      label: i18n.t('main_menu.edit.name'),
+      submenu: [
+        {
+          label: i18n.t('main_menu.edit.undo'),
+          accelerator: 'CmdOrCtrl+Z',
+          role: 'undo'
+        },
+        {
+          label: i18n.t('main_menu.edit.redo'),
+          accelerator: 'Shift+CmdOrCtrl+Z',
+          role: 'redo'
+        },
+        {
+          type: 'separator'
+        },
+        {
+          label: i18n.t('main_menu.edit.cut'),
+          accelerator: 'CmdOrCtrl+X',
+          role: 'cut'
+        },
+        {
+          label: i18n.t('main_menu.edit.copy'),
+          accelerator: 'CmdOrCtrl+C',
+          role: 'copy'
+        },
+        {
+          label: i18n.t('main_menu.edit.paste'),
+          accelerator: 'CmdOrCtrl+V',
+          role: 'paste'
+        },
+        {
+          label: i18n.t('main_menu.edit.select_all'),
+          accelerator: 'CmdOrCtrl+A',
+          role: 'selectall'
+        }
+      ]
+    },
+    {
+      label: i18n.t('main_menu.view.name'),
+      submenu: [
+        {
+          label: i18n.t('main_menu.view.toggle_full_screen'),
+          role: 'togglefullscreen'
+        }
+      ]
+    },
+    {
+      label: i18n.t('main_menu.window.name'),
+      submenu: [
+        {
+          label: i18n.t('main_menu.window.close'),
+          role: 'close'
+        },
+        {
+          label: i18n.t('main_menu.window.minimize'),
+          role: 'minimize'
+        },
+        {
+          type: 'separator'
+        },
+        {
+          label: i18n.t('main_menu.window.jump_to'),
+          accelerator: 'CmdOrCtrl+K',
+          enabled: true,
+          click: () => {
+            mainWindow.webContents.send('CmdOrCtrl+K')
+          }
+        },
+        {
+          type: 'separator'
+        },
+        ...accountsChange
+      ]
+    }
+  ]
+
+  const menu = Menu.buildFromTemplate(template)
+  Menu.setApplicationMenu(menu)
+}
